Db creation

+++ app/schemas.py
@@ -1,13 +1,15 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field, field_validator
 
 
 class OperationRequest(BaseModel):
     wallet_name: str = Field(..., max_length=127)
-    amount:float
+    amount: Decimal
     description: str | None = Field(None, max_length=255)
 
     @field_validator('amount')
-    def amount_must_be_positive(cls, value: float) -> float:
+    def amount_must_be_positive(cls, value: Decimal) -> Decimal:
         if value <= 0:
             raise ValueError("Amount must be positive")
         
@@ -25,7 +27,7 @@ class OperationRequest(BaseModel):
 
 class CreateWalletRequest(BaseModel):
     name: str = Field(..., max_length=127)
-    initial_balance: float = 0
+    initial_balance: Decimal = 0
 
     @field_validator("name")
     def name_not_empty(cls, value: str) -> str:
@@ -37,7 +39,7 @@ class CreateWalletRequest(BaseModel):
         return value
     
     @field_validator("initial_balance")
-    def balance_not_negative(cls, value: float) -> float:
+    def balance_not_negative(cls, value: Decimal) -> Decimal:
         if value < 0:
             raise ValueError("Initial balance cannot be negative")
